wip - table asks store for updated and store events updated the table, also when loading from  a node page: node arrived event is handeled by nodes.* event and not node update (Nodes component still determines selectedNodeId in a node load page).

+++ app/components/nodes/NodesList.js
@@ -20,13 +20,19 @@ export default class NodesList extends Box {
       this.renderLoading(diff.rhs);
     });
     
-    nodesStore.on('nodes', () => {
-      if (!this.tableChangedNodes) {
-        $(`#${this.componentId} table`).DataTable().ajax.reload();
+    nodesStore.on('nodesList.nodes.*', () => {
+      this.tabelDraw({
+        data: this._renderServerNodes(nodesStore.nodes),
+        draw: nodesStore.state.nodesList.metaData.draw,
+        recordsTotal: nodesStore.state.nodesList.metaData.recordsTotal,
+        recordsFiltered: nodesStore.state.nodesList.metaData.recordsFiltered
+      });
+      if (nodesStore.state.nodesList.search) {
+        $(`#${this.componentId} .dataTables_filter .input-group input`).val(nodesStore.state.nodesList.search);
       }
     });
 
-    nodesStore.on('nodes.*.timeSinceSync', (diff) => {
+    nodesStore.on('nodesList.nodes.*.timeSinceSync', (diff) => {
       var nodeIndex = diff.path[1];
       $(`#${this.componentId} tr input[data-index=${nodeIndex}]`)
         .parent().parent().find('td:nth-child(8n+8)') // last seen td
@@ -38,7 +44,19 @@ export default class NodesList extends Box {
   }
 
   _handlers() {
-    let _this = this;
+    var order = nodesStore.state.nodesList.order;
+    var orderDir = nodesStore.state.nodesList.orderDir;
+    var columns = [
+      { data: 'checkbox' },
+      { data: 'warning' },
+      { data: 'name', name: 'name' },
+      { data: 'tags', name: 'tags' },
+      { data: 'id', name: 'id' },
+      { data: 'platform', name: 'info.platform' },
+      { data: 'agentVersion', name: 'info.agentVersion' },
+      { data: 'lastSeen', name: 'lastSync' },
+      { data: 'deleteButton'}
+    ];
 
     $(`#${this.componentId} table`).DataTable({
       "pagingType": "full_numbers",
@@ -46,75 +64,41 @@ export default class NodesList extends Box {
       "<'row'<'col-sm-12'tr>>" +
       "<'row'<'col-sm-5'i><'col-sm-7'p>>",
       "oLanguage": { "sSearch": '' },
-      "serverSide": true,
-      "ajax": {
-        "url": '/api/nodes',
-        "data" : (data) => {
-          // if fetching data from a node page (loading a node page as first)
-          data.search.value = data.search.value || routerStore.nodeId();
-        },
-        'beforeSend': function (request) {
-          // customize request to server
-          request.setRequestHeader("X-CSRF-Token", window.csrfToken);
-        },
-        "dataSrc": (json) => {
-          // customize the server response and render the nodes
-          this.nodesToSet = json.nodes;
-          return this._renderServerNodes(json);
-        }
-      },
-      "order": [[ 7, "desc" ]],
+      "columns": columns,
+      "order": [[ columns.findIndex(c => c.name == order), orderDir ]],
       "columnDefs": [{ "targets": [0,1,3,4,8], "sortable": false }],
-      "columns"    : [
-        { data: 'checkbox' },
-        { data: 'warning' },
-        { data: 'name', name: 'name' },
-        { data: 'tags', name: 'tags' },
-        { data: 'id', name: 'id' },
-        { data: 'platform', name: 'info.platform' },
-        { data: 'agentVersion', name: 'info.agentVersion' },
-        { data: 'lastSeen', name: 'lastSync' },
-        { data: 'deleteButton'}
-      ]
-    }).on('preXhr', () => { this.renderLoading(true) })
-      .on('preDrawCallback', () => {
-        this._tableHandlersOff();
-      })
-      .on('draw.dt', () => {
-        this.tableChangedNodes = true;
-        nodesStore.setNodes(this.nodesToSet);
-        this.tableChangedNodes = false;
-        nodesStore.uncheckAllNodes();
-        $(`#${this.componentId} input:checkbox`).iCheck('uncheck');
-        this._tableHandlers();
-        this.renderLoading(false);
-      });
+      "pageLength": nodesStore.state.nodesList.length,
+      "serverSide": true,
+      "ajax" : this._fetchTableData.bind(this),
+    }).on('preDrawCallback', this._tableHandlersOff.bind(this))
+      .on('draw.dt', this._tableHandlers.bind(this));
 
-    // send table ajax search only on enter key and not on every key
+
+    // search table handler
     $(`#${this.componentId} .dataTables_filter input`).off().keyup(e => {
       var val = $(e.target).val();
       if (e.keyCode == 13 || (e.keyCode == 8 && !val)) {
-        this._tableSearch($(e.target).val());
+        nodesStore.fetchNodes({ search: val });
       }
     });
 
     // append 'actions' button to table
     $('.dataTables_filter > label').addClass('input-group').prepend(actionsTemplate);
-
     $(`#${this.componentId} a[name='aUpdateNode']`).click(() => {
       this.updateNodesVersionDialog.show();
     });
 
+    // check / uncheck all handler
     $(`#${this.componentId} input:checkbox[name='checkAllNodes']`).on('change', elem => {
-      var checkedIndexes = [];
+      var checkedNodes = [];
       if (elem.target.checked) {
         $(`#${this.componentId} input:checkbox[name='checkNode']`).iCheck('check').each((i, elem) =>  {
-          checkedIndexes.push(parseInt($(elem).data('index')));
+          checkedNodes.push($(elem).data('id'));
         });
-        nodesStore.setCheckedIndexes(checkedIndexes);
+        nodesStore.setCheckedNodeIds(checkedNodes);
       } else {
         $(`#${this.componentId} input:checkbox[name='checkNode']`).iCheck('uncheck');
-        nodesStore.uncheckAllNodes();
+          nodesStore.uncheckAllNodes();
       }
     });
   }
@@ -136,13 +120,10 @@ export default class NodesList extends Box {
 
     var query = routerStore.urlValueOf('q');
     if (query) {
-      this._tableSearch(routerStore.urlValueOf('q'));
+      nodesStore.fetchNodes({ search: routerStore.urlValueOf('q') });
     } else {
       this._focusTableSearch();
     }
-
-    // the table turns this off on draw event
-    this.renderLoading(true);
   }
 
   initialView() {
@@ -158,33 +139,44 @@ export default class NodesList extends Box {
     this._focusTableSearch();
   }
 
+  _nodeById(id) {
+    var node = nodesStore.nodes[id];
+    if (node) {
+      return node;
+    } else {
+      window.console.log(`sorry cannot find node id = ${id}`);
+      return null;
+    }
+  }
+
   _focusTableSearch() {
     $(`#${this.componentId} .dataTables_filter .input-group input`).focus();
   }
 
-  _tableSearch(value) {
-    $(`#${this.componentId} table`).DataTable().search(value).draw();
-  }
-
   _tableHandlers() {
-    $(`#${this.componentId} a[name='btSelectItemNodes']`).click(event => {
-      var index = parseInt($(event.target).data('index'));
-      var node = nodesStore.state.nodes[index];
-      routerStore.changeRoute('/nodes#' + node.id, node.name);
-    });
-
-    $(`#${this.componentId} input:checkbox[name='checkNode']`).on('change', event => {
-      nodesStore.toggleNode(parseInt($(event.target).data('index')));
-    });
-
     var _this = this;
+    $(`#${this.componentId} a[name='btSelectItemNodes']`).click(function(event) {
+      var node = _this._nodeById($(this).data('id'));
+      if (node) {
+        routerStore.changeRoute('/nodes#' + node.id, node.name);
+      }
+    });
+
+    $(`#${this.componentId} input:checkbox[name='checkNode']`).on('change', function(event) {
+      var node = _this._nodeById($(this).data('id'));
+      if (node) {
+        nodesStore.toggleNode(node.id);
+      }
+    });
+
     $(`#${this.componentId} button[name='btRemoveNode']`).click(function(event) {
-      let index = parseInt($(this).data('index'));
-      let node = nodesStore.state.nodes[index];
-      _this.confirmDialog.show({
-        ok: ()=> { nodesStore.deleteNode(node.id) },
-        text: `Remove node "${node.name}"?`
-      });
+      var node = _this._nodeById($(this).data('id'));
+      if (node) {
+        _this.confirmDialog.show({
+          ok: ()=> { nodesStore.deleteNode(node.id) },
+          text: `Remove node "${node.name}"?`
+        });
+      }
     });
 
     $(`#${this.componentId} table [data-toggle="popover"]`).popover();
@@ -196,24 +188,37 @@ export default class NodesList extends Box {
     $(`#${this.componentId} input:checkbox[name='checkNode']`).off();
     $(`#${this.componentId} button[name='btRemoveNode']`).off();
   }
+  
+  _fetchTableData(data, cb, settings) {
+    var nodesMetaData = {
+      start: data.start,
+      length: data.length,
+      order: data.columns[data.order[0].column].name,
+      orderDir: data.order[0].dir,
+      search: data.search.value || routerStore.nodeId()
+    };
+    nodesStore.fetchNodes(nodesMetaData);
+    this.tabelDraw = cb;
+  }
 
-  // this function receives the returned json from the nodes api and returns render-able data for the DataTable
-  _renderServerNodes(json) {
-    var renderedItems = json.nodes.map((node, i) => {
-      var node = nodesStore.enrich(node);
+  // this function receives nodes and returns render-able data for the DataTable
+  _renderServerNodes(nodes) {
+    var node;
+    var data = Object.keys(nodes).map(nodeId => {
+      node = nodes[nodeId];
       return {
-        checkbox: checkboxTemplate({index: i}),
+        checkbox: checkboxTemplate({id: nodeId}),
         warning: warningTemplate({node: node}),
-        name: nodeName({node: node, index: i}) || '',
+        name: nodeName({node: node}) || '',
         tags: tags({node: node}) || '',
-        id: node.id || '',
+        id: nodeId,
         platform: node.info.platform || '',
         agentVersion: node.info.agentVersion || '',
         lastSeen: node.timeSinceSync || '',
-        deleteButton: removeButtonTemplate({index: i})
+        deleteButton: removeButtonTemplate({id: nodeId})
       };
     });
 
-    return renderedItems;
+    return data;
   }
 }
